+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Changed
+- Filter out offline peers for data resolving #947
 
 ## [v2.0.1-rc] - 2020-03-17
-### Added
+### Changed
 - Instantiate CloudStorage for Genesis flow
 
 ## [v2.0.0-rc] - 2020-03-17